0.6: icon system — embedded Lucide SVGs via IconLibrary; replace all placeholder glyphs
- Add ui/IconLibrary.{hpp,cpp}: cached, theme-tinted juce::Drawable rendering
  from BinaryData SVGs (hathor_icon_data target); Lucide ISC licence at
  ui/resources/icons/LICENSE.
- ActivityRibbon letter labels (E/S/V/D/>/!/A/P) -> real icons.
- ChatSidebar U+F00D close glyph -> Lucide x IconButton.
- BreadcrumbsBar emoji (zap/search/columns) -> icons.
- StatusRibbon painted dots/shapes -> warning/play/stop/activity/git icons.
- ExplorerTreeItems inline geometry glyphs -> folder/music/code/wave/file icons.
All icons tint from the active palette and recolour on theme switch.

+++ ui/ActivityRibbon.cpp
@@ -105,7 +105,7 @@ void ActivityRibbon::paintButton(juce::Graphics& g, const RibbonButton& btn) con
 
     constexpr int kIconSize = 18;
     IconLibrary::drawIcon(g, btn.icon,
-                          btn.bounds.withSizeKeepingCentre(kIconSize, kIconSize),
+                          btn.bounds.withSizeKeepingCentre(kIconSize, kIconSize).toFloat(),
                           iconCol);
 }
